+++ src/env.ts
@@ -1,6 +1,6 @@
 import { Config, Options, parseValue, Result, Results } from "./common"
 
-export function parseEnvironment<O extends Options>(cfg: Config<O>): Results<O> & {_: string[]} {
+export function parseEnvironment<O extends Options>(cfg: Config<O>): Results<O> {
 
   const args = (cfg as any)._mockEnv || process.env
   const results: {[option: string]: Result} = {}