+++ test/index.js
@@ -1,4 +1,3 @@
-// <referance path="./types.d.ts" />
 const {parse} = require('../lib/index.js')
 const assert = require('assert')
 
@@ -54,24 +53,27 @@ assert.deepEqual(
   },
 )
 
-// require('./config.js') // TODO
+require('./config.js')
 require('./env.js')
 require('./args.js')
 
-console.log('Arguments overwrite environment')
+console.log('Arguments overwrite environment and config')
 assert.deepEqual(
   parse({
     options: {
       STR_OPT: {type: 'string'},
       NUM_OPT: {type: 'number', short: 'n'},
+      LIST_OPT: {type: 'list'},
     },
-    _mockEnv: {STR_OPT: 'a', NUM_OPT: '1'},
+    _mockEnv: {STR_OPT: 'a', NUM_OPT: '1', LIST_OPT: 'x'},
     _mockArgs: ['--num-opt', '2'],
+    _mockConfig: {list_opt: ['0']},
   }),
   {
     _: [],
     STR_OPT: 'a',
     NUM_OPT: '2',
+    LIST_OPT: ['x'],
   },
 )